Update feature-exception-handling from master

+++ src/main/java/com/is/countryneighborstour/services/ExchangeRateService.java
@@ -1,7 +1,8 @@
 package com.is.countryneighborstour.services;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface ExchangeRateService {
-    void saveExchangeRates(String baseCurrency, Map<String, Double> rates);
+    void saveExchangeRates(String baseCurrency, Map<String, BigDecimal> rates);
 }
